extract base base Platform methods
